在原有调查问卷功能的基础上，保留“项目基本信息”、“问卷基本信息”，删除其他板块，添加调查问卷（图片，多个文件上传）、访谈表（图片，多个文件上传）、调查现场（图片，多个文件上传）的上传

+++ SaPlat-service/SaPlat-service-provider/src/main/java/io/jboot/admin/service/provider/ManagementServiceImpl.java
@@ -40,7 +40,7 @@ public class ManagementServiceImpl extends JbootServiceBase<Management> implemen
      * @return
      */
     public Management fitModel(Management model) {
-        if (null != model && model.getSuperiorID() > 0) {
+        if (null != model && model.getSuperiorID().intValue() > 0) {
             Management tmp = findById(model.getSuperiorID());
             if (tmp != null) {
                 model.setSuperiorName(tmp.getName());
